Relocated the debug and persistence variables to the top of the script for easier access and configuration.
Introduced a new `DEBUG` Boolean variable, which when set to `True` would print out debug statements. The status messages such as `keypath` and enumerated `found_docs` hence, will only be printed when DEBUG is set to `True`.

Removed superfluous comments and also reorganized some segments of the code for better readability and brevity.